+++ fla/models/linear_attn/configuration_linear_attn.py
@@ -15,7 +15,7 @@ class LinearAttentionConfig(PretrainedConfig):
         attn_mode: str = "fused_chunk",
         hidden_size: int = 2048,
         expand_k: float = 1.0,
-        expand_v: int = 1,
+        expand_v: float = 1.0,
         hidden_ratio: Optional[int] = 4,
         intermediate_size: Optional[int] = None,
         num_hidden_layers: int = 24,
